Start work on versioned configs.
